# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CLinic Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/chronics/diseases": {
            "get": {
                "tags": [
                    "Chronic Diseases"
                ],
                "summary": "Get all chronic diseases for a patient",
                "operationId": "5c534e5e7e7a566ba5c2913815f89994",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "query",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chronic diseases",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChronicDiseasResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Chronic Diseases"
                ],
                "summary": "Create a new chronic disease",
                "operationId": "01311df30adb996e87e20f58f474becc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChronicDiseaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chronic disease created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChronicDiseasResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/chronics/diseases/{chronicDisease}": {
            "put": {
                "tags": [
                    "Chronic Diseases"
                ],
                "summary": "Update a chronic disease",
                "operationId": "460202695902e5d760959b266a109304",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chronic disease to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChronicDiseaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chronic disease updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChronicDiseasResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chronic disease not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Chronic Diseases"
                ],
                "summary": "Delete a chronic disease",
                "operationId": "c2884877017d0420b4247bf71603ae00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chronic disease to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Chronic disease not found"
                    }
                }
            }
        },
        "/api/chronics/medications": {
            "get": {
                "tags": [
                    "Chronic Medications"
                ],
                "summary": "Get all chronic medications for a patient",
                "operationId": "f3f89be00b34130d4ed54ca10e0b1a41",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "query",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chronic medications",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChronicMedicationResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Chronic Medications"
                ],
                "summary": "Create a new chronic medication",
                "operationId": "598917d63b208df7a45a3c5e55171773",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChronicMedicationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chronic medication created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChronicMedicationResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/chronics/medications/{chronicMedication}": {
            "put": {
                "tags": [
                    "Chronic Medications"
                ],
                "summary": "Update a chronic medication",
                "operationId": "ba803953963e995de9599e5d9916bc2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chronic medication to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChronicMedicationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chronic medication updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChronicMedicationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chronic medication not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Chronic Medications"
                ],
                "summary": "Delete a chronic medication",
                "operationId": "8c04cb44c22fea10e061353a0fe13389",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chronic medication to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Chronic medication not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChronicDiseaseRequest": {
                "required": [
                    "patient_id",
                    "description"
                ],
                "properties": {
                    "patient_id": {
                        "description": "The ID of the patient",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "description": "Description of the chronic disease",
                        "type": "string",
                        "example": "Diabetes"
                    }
                },
                "type": "object"
            },
            "ChronicMedicationRequest": {
                "title": "ChronicMedicationRequest",
                "description": "Request schema for chronic medication",
                "required": [
                    "patient_id",
                    "description"
                ],
                "properties": {
                    "patient_id": {
                        "description": "The ID of the patient",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "description": {
                        "description": "Description of the chronic medication",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Daily medication for hypertension"
                    }
                },
                "type": "object"
            },
...
            "ChronicDiseasResource": {
                "title": "Chronic Disease Resource",
                "description": "Chronic Disease resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the chronic disease",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "description": {
                        "description": "The description of the chronic disease",
                        "type": "string",
                        "example": "Diabetes"
                    }
                },
                "type": "object"
            },
            "ChronicMedicationResource": {
                "title": "Chronic Medciation Resource",
                "description": "Chronic Disease resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the chronic medciation",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "description": {
                        "description": "The description of the chronic disease",
                        "type": "string",
                        "example": "Diabetes"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}